rename ts files into PascalCase
