+++ src/FastCache.ts
